+++ planckton/sim.py
@@ -7,7 +7,7 @@ import hoomd.md
 from mbuild.formats.hoomd_simulation import create_hoomd_simulation
 import unyt as u
 
-from planckton.utils.utils import set_coeffs
+from planckton.utils.solvate import set_coeffs
 
 
 class Simulation:
